move realm init
