+++ app/src/main/java/org/macho/beforeandafter/MainActivity.kt
@@ -31,27 +31,11 @@ class MainActivity: AppCompatActivity() {
 
         colorSelected = ContextCompat.getColor(this, R.color.colorBottomNaviItemPressed)
 
-        configureRealm()
-
         uncheckCurrentButtonAndCheck(item0ImageButton, item0TextView)
 
         supportFragmentManager.beginTransaction().add(R.id.content, RecordFragment.getInstance()).commit()
 
         configureAd()
-    }
-
-    private fun configureRealm() {
-        Realm.init(applicationContext)
-        val config = RealmConfiguration.Builder()
-                .schemaVersion(1)
-                .migration { realm, oldVersion, newVersion ->
-                    val schema = realm.schema
-                    if (oldVersion == 0L) {
-                        schema.get("RecordDto").addField("memo", String::class.java)
-                    }
-                }
-                .build()
-        Realm.setDefaultConfiguration(config)
     }
 
     private fun uncheckCurrentButtonAndCheck(imageButton: ImageButton, textView: TextView) {
